Fix project images
The project images could not be loaded when under the `src` directory
because of the way webpack packages the files for distribution and the
way that the images were loaded in the `Project` component. Since we
need to load the image through its URL, it needs to be part of the
`public` directory which will automatically be included in the built
distribution, as opposed to the content of the `src` directory for which
only the files that have been referenced by your JSX code will be
bundled.