Fix menu display toggle
There are three main elements in this fix. The first one is the toggle
action to call `setMenuOpen`. Since it needs to be done on the `button`
element of the `Burger` component, I think it makes more sense to pass
the reference to the state value and setter so that the `Burger` can
update the state of its parent `Menu`. It's then the `Burger`'s
responsibility to call `setMenuOpen` with the right value when clicked.

Then comes the second part of the fix. In order to leverage animations
and style overrides in CSS, it's better to use classes rather than
styles directly. One of the reasons why I find it easier to work with
classes is that with a combination of classes you can override some
style definitions, whereas the `style` attribute of a component
overrides all other values. Once both classes are defined (I called them
`menuContainer` and `menuContainer__Visible` in this case), we can
leverage the `menuOpen` state to either assign one class or both of them
to the `aside` element. To me, it's the easiest way to leverage
Javascript to toggle styles, leaving all the CSS definitions in distinct
classes.